requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anchorpy==0.10.0
anyio==4.8.0
apischema==0.17.5
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
base58==2.1.1
based58==0.1.1
bip_utils==2.7.0
bitarray==2.9.3
bitcoinlib==0.7.1
borsh-construct==0.1.0
bunch==1.0.1
CacheControl==0.14.2
cached-property==2.0.1
cachetools==4.2.4
cbor2==5.6.5
certifi==2024.12.14
cfscrape==2.1.1
channels==4.0.0
channels_redis==4.2.1
charset-normalizer==3.4.1
ckzg==1.0.2
click==8.1.8
coincurve==17.0.0
constantly==23.10.4
construct==2.10.68
construct-typing==0.5.6
crcmod==1.7
cryptography==41.0.7
cytoolz==0.12.3
daphne==4.1.2
Django==4.2.18
django-cors-headers==4.3.1
django-jazzmin==3.0.1
django-jet-reboot==1.3.10
django-redis==5.4.0
djangorestframework==3.14.0
ecdsa==0.19.0
ed25519-blake2b==1.4.1
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.7.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==3.0.0
eth-utils==1.9.5
exceptiongroup==1.2.2
fastecdsa==3.0.1
filelock==3.17.0
frozendict==2.3.10
frozenlist==1.5.0
h11==0.14.0
hdwallet==2.2.1
hexbytes==0.3.1
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
jsonalias==0.1.1
jsonrpcclient==4.0.3
jsonrpcserver==5.0.9
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lockfile==0.12.2
lru-dict==1.2.0
mnemonic==0.20
moralis==0.1.49
more-itertools==8.14.0
msgpack==1.1.0
multiaddr==0.0.9
multidict==6.1.0
mysqlclient==2.2.7
nest-asyncio==1.6.0
netaddr==1.3.0
numpy==2.0.2
OSlash==0.6.3
packaging==24.2
parsimonious==0.8.1
pillow==10.2.0
pluggy==1.5.0
propcache==0.2.1
protobuf==3.19.5
psutil==6.1.1
py==1.11.0
py-ecc<7.0.0
py-sr25519-bindings==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pyheck==0.1.5
pyOpenSSL==25.0.0
pytest==6.2.5
pytest-asyncio==0.17.2
pytest-xprocess==0.18.1
python-bitcoinlib==0.12.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python_avatars==1.4.1
pytz==2024.2
pyunormalize==16.0.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
rfc3986==1.5.0
rlp==2.0.1
rpds-py==0.22.3
service-identity==24.2.0
simplejson==3.19.3
sniffio==1.3.1
solana==0.25.1
solana-sdk==0.25.6
solders==0.2.0
spl==0.0.1.dev0
SQLAlchemy==2.0.37
sqlparse==0.5.3
sumtypes==0.1a6
toml==0.10.2
tomli==2.2.1
toolz==0.11.2
Twisted==24.11.0
txaio==23.1.1
types-cachetools==4.2.10
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
varint==1.0.2
web3==5.31.4
websockets>=10.3,<11.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.17.0
openai>=1.0.0